requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-debug-toolbar==4.3.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
flake8==7.0.0
idna==3.6
isort==5.13.2
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
stripe==8.4.0
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.12
typing_extensions==4.10.0
urllib3==2.2.1
